feat: merge actions into create or update contact
