+++ docs/_snippets/apps.mdx
@@ -323,6 +323,15 @@
       </Card>
     </a>
       
+    <a href="https://activepieces.com/docs/pieces/apps/sendinblue">
+      <Card>
+          <p align="center">
+            <strong>Sendinblue</strong>
+            <img height="75px" width="75px" src="https://cdn.activepieces.com/pieces/sendinblue.png" />
+          </p>
+      </Card>
+    </a>
+      
     <a href="https://activepieces.com/docs/pieces/apps/slack">
       <Card>
           <p align="center">
