+++ docs/mint.json
@@ -100,6 +100,7 @@
         "pieces/apps/rss",
         "pieces/apps/schedule",
         "pieces/apps/sendgrid",
+        "pieces/apps/sendinblue",
         "pieces/apps/slack",
         "pieces/apps/square",
         "pieces/apps/store",
