+++ docs/pieces/apps/sendinblue.mdx
@@ -3,7 +3,9 @@ title: 'Sendinblue'
 ---
 
 ## Triggers
-  
+
+*No supported triggers yet, please let us know if you need something on Discord so we can help out* 
+
 
 ---
 
@@ -11,14 +13,8 @@ title: 'Sendinblue'
 
 
     <CardGroup cols={2}>
-      <Card title="Create Contact">
-        Create a new Contact
-      </Card>
-    </CardGroup>
-  
-    <CardGroup cols={2}>
-      <Card title="Update Contact">
-        Update an existing contact
+      <Card title="Create or Update Contact">
+        Create or update an existing contact
       </Card>
     </CardGroup>
